% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMM_proj.R
\name{ADMM_proj}
\alias{ADMM_proj}
\title{ADMM algorithm}
\usage{
ADMM_proj(
  mat,
  epsilon = 1e-04,
  mu = 10,
  it.max = 1000,
  etol = 1e-04,
  etol_distance = 1e-04
)
}
\arguments{
\item{mat}{Matrix to be projected}

\item{epsilon}{Approximation of the space of positive semi-definite matrix at epsilon}

\item{mu}{Penalty parameter of ADMM algorithm}

\item{it.max}{Number maximum of iterations}

\item{etol}{Tolerance parameter for the convergence of primal and dual residual}

\item{etol_distance}{Tolerance parameter for the convergence of the distance}
}
\value{
list containing \itemize{
\item \code{mat} Projected matrix
\item \code{df_ADMM} dataframe containing parameters of the ADMM algorithm for each iteration of the algorithm
}
}
\description{
Finds the nearest positive semi-definite matrix with respect to the max norm
}
\seealso{
\url{https://web.stanford.edu/~boyd/papers/pdf/admm_distr_stats.pdf}
}
