% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockwise_coordinate_descent.R
\name{blockwise_coordinate_descent}
\alias{blockwise_coordinate_descent}
\title{Blockwise coordinate descent}
\usage{
blockwise_coordinate_descent(Z, y, n, p, p1, p2,
  lambda.factor = ifelse(dim(Z)[1] < dim(Z)[2], 0.01, 0.001),
  step = 100, K = 4, mu = 10, tau = NULL, ratio_matrix = NULL,
  etol = 1e-04, optTol = 1e-05, earlyStopping_max = 10,
  noise = c("additive", "missing", "HM"))
}
\arguments{
\item{Z}{Corrupted design matrix (with additive error or missing data)}

\item{y}{Response vector}

\item{n}{Number of samples of the design matrix}

\item{p}{Number of features of the matrix}

\item{p1}{Number of uncorrupted predictors}

\item{p2}{Number of corrupted predictors}

\item{lambda.factor}{Range of the lambda interval we are going to explore}

\item{step}{Number of values of lambda in the interval we are going to test}

\item{K}{Number of folds for the cross-validation}

\item{mu}{Penalty parameter for the ADMM algorithm}

\item{tau}{Standard deviation for the additive error matrix in the additive error setting (NULL in the missing data setting)}

\item{ratio_matrix}{Observation matrix used in the missing data setting}

\item{etol}{Tolerance parameter for the ADMM algorithm}

\item{optTol}{Tolerance parameter for the convergence of the error in the pathwise coordinate descent}

\item{earlyStopping_max}{Number of iterations allowed when error starts increasing}

\item{noise}{Type of noise (additive or missing)}
}
\value{
list containing \itemize{
\item \code{lambda.opt} optimal value of lambda corresponding to minimum error
\item \code{lambda.sd} Value of lambda corresponding to error higher than minimum error by one standard deviation
\item \code{beta.opt} Value of beta corresponding to \code{lambda.opt}
\item \code{beta.sd} Value of beta corresponding to \code{lambda.sd}
\item \code{data_error} Dataframe containing errors and their standard deviation for each iteration of the algorithm
\item \code{data_beta} Dataframe containing the values of beta for each iteration of the algorithm
\item \code{earlyStopping} Integer containing the value of iteration when early stopping happens
}
}
\description{
Implement blockwise coordinate descent algorithm. Best penalty value is evaluated through cross-validation.
}
\examples{

}
