% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.coco}
\alias{coef.coco}
\title{Coef Method for \code{coco} object}
\usage{
coef.coco(object, s = NULL, ...)
}
\arguments{
\item{object}{Fitted \code{coco} model object}

\item{s}{Value(s) of the penalty parameter \code{lambda} at which predictions
are required. Default is the entire sequence used to create the model.}

\item{...}{currently ignored}
}
\value{
Coefficient value
}
\description{
Similar to other coef functions, this function returns the coefficient values 
for different possible lambda values
}
