% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{cov_autoregressive}
\alias{cov_autoregressive}
\title{Autoregressive covariance matrix}
\usage{
cov_autoregressive(p)
}
\arguments{
\item{p}{Number of features}
}
\value{
Covariance matrix
}
\description{
Create autoregressive covariance matrix
}
