% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_covariance_matrices.R,
%   R/oxy-cv_covariance_matrices.R
\name{cv_covariance_matrices}
\alias{cv_covariance_matrices}
\title{Projected covariance matrices}
\usage{
cv_covariance_matrices(K, mat, y, p, mu, tau = NULL,
  ratio_matrix = NULL, etol = 1e-04, noise = c("additive", "missing",
  "HM"))

cv_covariance_matrices(K, mat, y, p, mu, tau = NULL,
  ratio_matrix = NULL, etol = 1e-04, noise = c("additive", "missing",
  "HM"))
}
\arguments{
\item{K}{number of folds for the cross validation}

\item{mat}{covariance matrix to be projected}

\item{y}{response vector}

\item{p}{number of predictors}

\item{mu}{penalty parameter for the ADMM algorithm.}

\item{tau}{standard error of the additive error matrix when the chosen setting is the additive error setting}

\item{ratio_matrix}{observation matrix used in the missing data setting}

\item{etol}{PARAM_DESCRIPTION, Default: 1e-04}

\item{noise}{type of setting chosen : additive or missing}

\item{K}{PARAM_DESCRIPTION}

\item{mat}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}

\item{p}{PARAM_DESCRIPTION}

\item{mu}{PARAM_DESCRIPTION}

\item{tau}{PARAM_DESCRIPTION, Default: NULL}

\item{ratio_matrix}{PARAM_DESCRIPTION, Default: NULL}

\item{noise}{PARAM_DESCRIPTION, Default: c("additive", "missing", "HM")}
}
\value{
list containing \itemize{
\item \code{sigma_global} projected matrix for \code{mat}
\item \code{rho_global} rho parameter for \code{mat}
\item \code{list_matrices_lasso} list of the projected matrices for \code{mat} deprived of the k-fold during cross validation
\item \code{list_matrices_error} list of the projected matrices for the k-fold of \code{mat} 
\item \code{list_rho_lasso} list of the modified \code{rho} for \code{mat} deprived of the K-fold during cross validation
\item \code{list_rho_error} list of the modified \code{rho} for the k-fold of \code{mat} 
}

OUTPUT_DESCRIPTION
}
\description{
Creating projected nearest positive semi-definite covariance matrices for the cross validation step of the CoCoLasso

FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
