% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_covariance_matrices_block_descent.R
\name{cv_covariance_matrices_block_descent}
\alias{cv_covariance_matrices_block_descent}
\title{Projected covariance matrices for BD-CoCoLasso}
\usage{
cv_covariance_matrices_block_descent(K, mat, y, p, p1, p2, mu,
  tau = NULL, ratio_matrix = NULL, etol = 1e-04,
  noise = c("additive", "missing"))
}
\arguments{
\item{K}{Number of folds for the cross validation}

\item{mat}{Covariance matrix to be projected}

\item{y}{Response vector}

\item{p}{Number of predictors}

\item{p1}{Number of uncorrupted predictors}

\item{p2}{Number of corrupted predictors}

\item{mu}{Penalty parameter for the ADMM algorithm.}

\item{tau}{Standard error of the additive error matrix when the chosen setting is the additive error setting}

\item{ratio_matrix}{Observation matrix used in the missing data setting}

\item{etol}{Tolerance used in the ADMM algorithm}

\item{noise}{Type of setting chosen : additive or missing}
}
\value{
list containing \itemize{
\item \code{sigma_global} projected matrix for \code{mat}
\item \code{rho_global} rho parameter for \code{mat}
\item \code{list_matrices_lasso} list of the projected matrices for \code{mat} deprived of the k-fold during cross validation
\item \code{list_matrices_error} list of the projected matrices for the k-fold of \code{mat} 
\item \code{list_rho_lasso} list of the modified \code{rho} for \code{mat} deprived of the K-fold during cross validation
\item \code{list_rho_error} list of the modified \code{rho} for the k-fold of \code{mat} 
}
}
\description{
Creating projected nearest positive semi-definite covariance matrices for the cross validation step of the BD-CoCoLasso algorithm. In that case, the design matrix must be organized as follozs : uncorrupted features must be the first block of the matrix, and corrupted features must be the second block of the matrix.
}
