% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HM_proj.R
\name{HM_proj}
\alias{HM_proj}
\title{HM-lasso algorithm}
\usage{
HM_proj(
  sigmaHat,
  R = NULL,
  a = 1,
  iter_max = 1000,
  epsilon = 1e-04,
  mu = 10,
  tolerance = 1e-04,
  norm = "F"
)
}
\arguments{
\item{sigmaHat}{covariance matrix}

\item{R}{weight matrix}

\item{a}{hmlasso (1) or cocolasso (0)}

\item{iter_max}{Number maximum of iterations}

\item{epsilon}{Stopping criterion}

\item{mu}{Penalty parameter of ADMM algorithm}

\item{tolerance}{Tolerance parameter for the convergence of primal and dual residual}

\item{norm}{Frobenius or max norm}
}
\value{
list containing \itemize{
\item \code{Ak} Projected matrix
}
}
\description{
Finds the nearest positive semi-definite matrix with respect to the Frobenius norm
}
\examples{
 
M = matrix(-1,20,20)
mat_proj <- BDcocolasso::HM_proj(mat=M)

}
\seealso{
\url{https://arxiv.org/pdf/1811.00255.pdf}
}
