% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_covariance.R
\name{lasso_covariance}
\alias{lasso_covariance}
\title{Lasso in covariance form}
\usage{
lasso_covariance(n, p, lambda, control = list(maxIter = 1000, optTol =
  10^(-5), zeroThreshold = 10^(-6)), XX, Xy, beta.start)
}
\arguments{
\item{n}{Number of samples of the design matrix}

\item{p}{Number of features of the matrix}

\item{lambda}{penalty parameter}

\item{control}{Including control parameters : max of iterations, tolerance for the convergence of the error, zero threshold to put to zero small beta coefficients}

\item{XX}{Design matrix corresponding to \eqn{\frac{1}{n} X'X} or a modified version in the case of CoCoLasso}

\item{Xy}{Rho parameter corresponding to \eqn{\frac{1}{n} X'y} or a modified version in the case of CoCoLasso}

\item{beta.start}{Initial value of beta}
}
\value{
list containing \itemize{
\item coefficients : Coefficients corresponding to final beta after convergence of the algoritm
\item coef.list : Matrix of coefficients for beta for all iterations
\item num.it Number of iterations 
}
}
\description{
Solve the least squares loss with lasso penalty written in a form with the covariance matrix : \eqn{\frac{1}{2} \beta^{'} \Sigma \beta - \rho^{'} \beta + \lambda \|\beta\|_1}
}
