% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_covariance_block_general.R
\name{lasso_covariance_block_general}
\alias{lasso_covariance_block_general}
\title{Lasso in covariance form for the BD-CoCoLasso}
\usage{
lasso_covariance_block_general(
  n,
  p1,
  p2,
  p3,
  X1,
  Z2,
  Z3,
  y,
  sigma1,
  sigma2,
  sigma3,
  lambda,
  ratio_matrix = NULL,
  control = list(maxIter = 1000, optTol = 10^(-5), zeroThreshold = 10^(-6)),
  beta1.start,
  beta2.start,
  beta3.start,
  penalty = c("lasso", "SCAD")
)
}
\arguments{
\item{n}{Number of samples of the design matrix}

\item{p1}{Number of uncorrupted predictors}

\item{p2}{Number of corrupted predictors containing additive error}

\item{p3}{Number of corrupted predictors containing missingness}

\item{X1}{first block of the design matrix corresponding to uncorrupted features}

\item{Z2}{second block of the design matrix corresponding to corrupted features containing additive error}

\item{Z3}{third block of the design matrix corresponding to corrupted features containing missingness}

\item{y}{Response vector}

\item{sigma1}{Covariance matrix for X1 : \eqn{\frac{1}{n} X_1'X_1}. This parameter is automatically furnished in \link{blockwise_coordinate_descent}}

\item{sigma2}{Modified covariance matrix for Z2 through the CoCoLasso algorithm. This parameter is automatically furnished in \link{blockwise_coordinate_descent}}

\item{sigma3}{Modified covariance matrix for Z3 through the CoCoLasso algorithm. This parameter is automatically furnished in \link{blockwise_coordinate_descent}}

\item{lambda}{Penalty parameter}

\item{ratio_matrix}{Observation matrix in the missing data setting (NULL otherwise)}

\item{control}{Including control parameters : max of iterations, tolerance for the convergence of the error, zero threshold to put to zero small beta coefficients}

\item{beta1.start}{Initial value for the coefficients of uncorrupted features}

\item{beta2.start}{Initial value for the coefficients of corrupted features containing additive error}

\item{beta3.start}{Initial value for the coefficients of corrupted features containing missingness}

\item{penalty}{Type of penalty used : can be lasso penalty or SCAD penalty}
}
\value{
list containing \itemize{
\item coefficients.beta1 : Coefficients corresponding to final beta1 after convergence of the algoritm
\item coefficients.beta2 : Coefficients corresponding to final beta2 after convergence of the algoritm
\item num.it : Number of iterations of algorithm
}
}
\description{
Solve the least squares loss with lasso penalty written in a form with the covariance matrix : \eqn{\frac{1}{2} \beta^{'} \Sigma \beta - \rho^{'} \beta + \lambda \|\beta\|_1}
}
