% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoef.R
\name{plotCoef}
\alias{plotCoef}
\title{Plot solution path}
\usage{
plotCoef(object, linetype = "dashed", col = "black")
}
\arguments{
\item{object}{A fitted \code{cocolasso} object as produced by \code{pathwise_coordinate_descent()} or 
\code{blockwise_coordinate_descent()}}

\item{linetype}{Linetype for the lines corresponding to the lambda values. Default is "dashed"}

\item{col}{Colour for the lines corresponding to lambda values. Default is "black"}
}
\value{
A plot is produced and nothing is returned
}
\description{
Plot evolution of coefficients for the range of lambda values
}
