% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotError.R
\name{plotError}
\alias{plotError}
\title{Plot solution path}
\usage{
plotError(object, linetype = "dashed", col = "black",
  colLine = "red")
}
\arguments{
\item{object}{A fitted \code{cocolasso} object as produced by \code{pathwise_coordinate_descent()} or 
\code{blockwise_coordinate_descent()}}

\item{linetype}{Linetype for the lines corresponding to the lambda values. Default is "dashed"}

\item{col}{Colour for the lines corresponding to lambda values. Default is "black"}

\item{colLine}{Colour of the error line. Default is "red"}
}
\value{
A plot is produced and nothing is returned
}
\description{
Plot evolution of coefficients for the range of lambda values
}
