% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.coco}
\alias{predict.coco}
\title{Make predictions from a coco object}
\usage{
\method{predict}{coco}(
  object,
  newx,
  s = NULL,
  lambda.pred = NULL,
  type = c("response", "coefficients"),
  ...
)
}
\arguments{
\item{object}{Fitted \code{coco} model object}

\item{newx}{matrix of new values for \code{x} at which predictions are to be
made. Do not include the intercept (this function takes care of that). Must
be a matrix. This argument is not used for \code{type=c("coefficients")}. This 
matrix must have the same number of columns originally supplied to the \code{coco}
fitting function.}

\item{s}{Value(s) of the penalty parameter \code{lambda} at which predictions
are required. Default is the entire sequence used to create the model.}

\item{lambda.pred}{Value(s) of the penalty parameter \code{lambda} at which coefficients
are extracted to calculate the response based on the matrix of new values \code{newx}.
Default is \code{lambda.sd}.}

\item{type}{Type of prediction required. Type \code{"coefficients"} computes the coefficients 
at the requested values for \code{s}. Type \code{response} computes the response based 
on the covariates values in \code{newx}, for coefficients corresponding to \code{lambda} value
in \code{lambda.pred}}

\item{...}{currently ignored}
}
\value{
The object returned depends on type.
}
\description{
Similar to other predict methods, this functions predicts fitted
  values, logits, coefficients and more from a fitted \code{coco} object.
}
