% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_data}
\alias{simulated_data}
\title{Simulated data}
\format{this data frame has 10000 rows and the following 50 columns:
\describe{
  \item{y}{the response}
  \item{V}{Other features}
}}
\usage{
simulated_data
}
\description{
Because of the constraint on data privacy, we could not use UK Biobank data set as an example, so we simulated data with missing values.
}
\keyword{datasets}
