% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_data_additive_block}
\alias{simulated_data_additive_block}
\title{Simulated data additive block}
\format{this data frame has 10000 rows and the following 200 columns:
\describe{
  \item{y}{the response}
  \item{V}{180 first columns are uncorrupted covariates}
  \item{V}{20 last columns are the corrupted covariates with additive error}
}}
\usage{
simulated_data_additive_block
}
\description{
Simulated data y=X1 \* beta1 + Z2 \* beta2 + epsilon where beta1=c(3,2,0,0,1.5,0,...) 
qnd beta2 = c(0,...,1.5,0,0,2,3). X1 is uncorrupted while Z2 is corrupted with additive error
}
\keyword{datasets}
