% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_data_missing}
\alias{simulated_data_missing}
\title{Simulated data missing}
\format{this data frame has 10000 rows and the following 50 columns:
\describe{
  \item{y}{the response}
  \item{V}{Other features}
}}
\usage{
simulated_data_missing
}
\description{
Simulated data y=X * beta + epsilon where beta=c(3,2,0,0,1.5,0,...) and with missing values
}
\keyword{datasets}
