% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.coco}
\alias{print.coco}
\title{Print Method for \code{coco} object}
\usage{
\method{print}{coco}(object)
}
\arguments{
\item{x}{fitted \code{coco} object}

\item{...}{additional print arguments}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Print a summary of the \code{coco} path at each step along the
  path.
}
\seealso{
\code{\link{coco}}
}
